Adopt commit 37560d6a2 from master: Merge pull request #1164 from oimanuel/add_unsigned_atomics-drqs-101285412
commit 37560d6a239b964cf655f53d4fecb49dfd3d6fbe
Merge: 5e0646212 3317a03d5
Author: Jeffrey Mendelsohn <JMENDELSOHN4@bloomberg.net>
Date:   Thu Jun 15 11:45:24 2017 -0400

    Merge pull request #1164 from oimanuel/add_unsigned_atomics-drqs-101285412

    add unsigned atomics DRQS 101285412

commit 3317a03d592cba967d466e771e07da60351d4fc2
Author: Ofer Imanuel <oimanuel@bloomberg.net>
Date:   Thu Jun 15 11:43:57 2017 -0400

    Jeff comments #2

commit e7b4243fe2f68a28454c9e178ece10a0a1cd2b22
Author: Ofer Imanuel <oimanuel@bloomberg.net>
Date:   Wed Jun 14 15:11:25 2017 -0400

    Jeff comments #1

commit df2436582711b9a56d007c4b742fa423fa876ea4
Author: Ofer Imanuel <oimanuel@bloomberg.net>
Date:   Mon Jun 5 14:43:13 2017 -0400

    add unsigned to test cases 6, 7, 8 bsls_atomic.t

commit 2938045e9763c6d6a35d0c54a0e034a580870f3b
Author: Ofer Imanuel <oimanuel@bloomberg.net>
Date:   Mon Jun 5 14:08:47 2017 -0400

    add unsigned to test cases 2, 3, 4, 5 for bsls_atomic.t

commit 573b6564e43245fb7464c0273291efaae2bf9cc5
Author: Ofer Imanuel <oimanuel@bloomberg.net>
Date:   Mon Jun 5 11:48:07 2017 -0400

    cleanup some unrelated build warnings

commit f35a910b4042000d96564df8acd4be6783a60e27
Author: Ofer Imanuel <oimanuel@bloomberg.net>
Date:   Mon Jun 5 11:46:56 2017 -0400

    fix error

commit 23eedfc976e49ed3e236dbe591ccd86d4f29df11
Author: Ofer Imanuel <oimanuel@bloomberg.net>
Date:   Mon Jun 5 11:02:19 2017 -0400

    added unsigned to test cases 11, 12 bsls_atomicoperations

commit cc95c5ffd717e7b075885767cfd7e32a9a6ffe30
Author: Ofer Imanuel <oimanuel@bloomberg.net>
Date:   Fri Jun 2 17:51:13 2017 -0400

    add unsigned for test cases 8, 10, bsls_atomicoperations.t

commit 5acd60d91c4107a91e6a284838068a9a4ea9b2cc
Author: Ofer Imanuel <oimanuel@bloomberg.net>
Date:   Fri Jun 2 15:24:39 2017 -0400

    fix test case 6; add unsigned for test case 6, bsls_atomicoperations.t

commit 19e7114d628ba047e9a7df6597cf23d601d52d68
Author: Ofer Imanuel <oimanuel@bloomberg.net>
Date:   Fri Jun 2 14:07:14 2017 -0400

    fix test case 5; add unsigned for test case 5, bsls_atomicoperations.t

commit 3ff53bad018b5ca06c49658daad9c0fa613b0659
Author: Ofer Imanuel <oimanuel@bloomberg.net>
Date:   Fri Jun 2 11:23:05 2017 -0400

    add case 4 unsigned to bsls_atomicoperations.t

commit dbd493d47207b0defc95079a318445a15002b1cd
Author: Ofer Imanuel <oimanuel@bloomberg.net>
Date:   Thu Jun 1 19:09:37 2017 -0400

    added cases 2, 3 for uint, uint64

commit cb9f8eaa3c6552f9db8afb829bd004af836546c1
Author: Ofer Imanuel <oimanuel@bloomberg.net>
Date:   Thu Jun 1 16:32:12 2017 -0400

    fix issue to compile on ibm andsun

commit 81eb8f1b988ebc13abf4def90a2041495fa6e3e1
Author: Ofer Imanuel <oimanuel@bloomberg.net>
Date:   Thu Jun 1 12:56:04 2017 -0400

    partial bde_verify cleanup for the code - old and new

commit 2cf2a80a9e221fc666cd08d251734050f9d7009b
Author: Ofer Imanuel <oimanuel@bloomberg.net>
Date:   Tue May 30 17:39:30 2017 -0400

    add Uint Uint64 traits for the rest of the types; some bde_verify cleanup of bsls_atomic.h, bsls_atomic.t.cpp,  bsls_atomicoperations.t.cpp

commit e8d86c6dca4f4b249b38b83f858a7075224bd6ce
Author: Ofer Imanuel <oimanuel@bloomberg.net>
Date:   Tue May 30 12:25:49 2017 -0400

    cleanup warnings on teat drivers

commit 118dba3ff38e1b86bf69ebb2cbae0612706b08c0
Author: Ofer Imanuel <oimanuel@bloomberg.net>
Date:   Tue May 30 11:27:31 2017 -0400

    added implementation to Uint64

commit ee77c3ede7323fc734aa1d8372161ee23a931f4d
Author: Ofer Imanuel <oimanuel@bloomberg.net>
Date:   Fri May 26 17:13:33 2017 -0400

    work in progress

+++ groups/bsl/bsls/bsls_atomicoperations_x86_all_gcc.t.cpp
@@ -227,8 +227,10 @@ void test_atomics(bsls::Types::Int64 iterations) {
         totalReadTime += readerArgs[i].d_runtimeMs;
     }
     static const double MS_PER_SEC = 1000;
-    double readPerSec = (double)totalReadIter / totalReadTime * MS_PER_SEC;
-    double writePerSec = (double)iterations / args.d_runtimeMs * MS_PER_SEC;
+    double readPerSec = (double)totalReadIter /
+                        (double)totalReadTime * MS_PER_SEC;
+    double writePerSec = (double)iterations /
+                         (double)args.d_runtimeMs * MS_PER_SEC;
 
     printf(" %f" " / %f\n", readPerSec, writePerSec);
